Running modes
Add a switcher before the target
According to the documentation (which is not completely accurate on that matter) The first argument is a target, id est a name with no leading dash. We can use that feature to branch to different tools.
The former l3build.lua has become l3build-mode-normal.lua, with some trailing addenda
